requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
psycopg2-binary>=2.9.9
gunicorn>=21.2.0
python-decouple>=3.8
Pillow>=10.1.0
django-cors-headers>=4.3.1
resend>=0.8.0
